# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mini DMS API
  version: 1.0.0
  description: Private document management with asynchronous extraction and recoverable
    jobs.
paths:
...
  /api/v1/documents/{id}/:
    get:
      operationId: api_v1_documents_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
        required: true
      tags:
      - Documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
    patch:
      operationId: api_v1_documents_partial_update
      parameters:
      - in: header
        name: If-Match
        schema:
          type: string
        description: 'Current metadata revision, quoted: "1".'
        required: true
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
        required: true
      tags:
      - Documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMetadataRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMetadataRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: ''
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '409':
          description: State or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '412':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '428':
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
    delete:
      operationId: api_v1_documents_destroy
      parameters:
      - in: header
        name: If-Match
        schema:
          type: string
        description: 'Current metadata revision, quoted: "1".'
        required: true
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
        required: true
      tags:
      - Documents
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '409':
          description: State or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '412':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '428':
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
...
  /api/v1/documents/{id}/status/:
    get:
      operationId: api_v1_documents_status_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
        required: true
      tags:
      - Processing
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingStatus'
          description: ''
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
  /api/v1/documents/{id}/text/:
    get:
      operationId: api_v1_documents_text_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
        required: true
      tags:
      - Processing
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OCRText'
          description: ''
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            X-Request-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID shared with audit and job logs.
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        owner_id:
          type: integer
          readOnly: true
        filename:
          type: string
          readOnly: true
        mime_type:
          type: string
          readOnly: true
        size:
          type: integer
          readOnly: true
        sha256:
          type: string
          readOnly: true
        title:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        document_type:
          type: string
          readOnly: true
        tags:
          readOnly: true
        metadata:
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        revision:
          type: integer
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        attempts:
          type: integer
          readOnly: true
        processed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error_code:
          type: string
          readOnly: true
        error_message:
          type: string
          readOnly: true
        processing_info:
          readOnly: true
      required:
      - attempts
      - created_at
      - description
      - document_type
      - error_code
      - error_message
      - filename
      - id
      - metadata
      - mime_type
      - owner_id
      - processed_at
      - processing_info
      - revision
      - sha256
      - size
      - status
      - tags
      - title
      - updated_at
...
    OCRText:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        text:
          type: string
      required:
      - id
      - status
      - text
...
    PatchedMetadataRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          default: ''
          maxLength: 10000
        document_type:
          type: string
          default: ''
          maxLength: 64
        tags: {}
        metadata: {}
...
    StatusEnum:
      enum:
      - pending
      - processing
      - succeeded
      - failed
      type: string
      description: |-
        * `pending` - Pending
        * `processing` - Processing
        * `succeeded` - Succeeded
        * `failed` - Failed
...
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          - request_id
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              format: uuid
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT